Move upsert/delete producer key functions to `producer-key-readmodel-writer-sql`
