+++ packages/authorization-process/test/integrationUtils.ts
@@ -38,7 +38,10 @@ import {
   producerKeychainReadModelServiceBuilder,
   purposeReadModelServiceBuilder,
 } from "pagopa-interop-readmodel";
-import { upsertAgreement } from "pagopa-interop-readmodel/testUtils";
+import {
+  upsertAgreement,
+  upsertProducerJWKKey,
+} from "pagopa-interop-readmodel/testUtils";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { authorizationServiceBuilder } from "../src/services/authorizationService.js";
 import { config } from "../src/config/config.js";
@@ -141,7 +144,7 @@ export const addOneKey = async (key: ClientJWKKey): Promise<void> => {
 export const addOneProducerKey = async (key: ProducerJWKKey): Promise<void> => {
   await writeInReadmodel(key, producerKeys);
 
-  await producerJWKKeyReadModelServiceSQL.upsertProducerJWKKey(key, 0);
+  await upsertProducerJWKKey(readModelDB, key, 0);
 };
 
 export const addOneClient = async (client: Client): Promise<void> => {
