+++ packages/producer-key-readmodel-writer-sql/src/index.ts
@@ -8,21 +8,13 @@ import {
   unsafeBrandId,
 } from "pagopa-interop-models";
 import { match } from "ts-pattern";
-import {
-  makeDrizzleConnection,
-  producerJWKKeyReadModelServiceBuilder,
-} from "pagopa-interop-readmodel";
+import { makeDrizzleConnection } from "pagopa-interop-readmodel";
 import { handleMessageV2 } from "./producerKeyConsumerServiceV2.js";
 import { config } from "./config/config.js";
-import { customReadModelServiceBuilder } from "./readModelService.js";
+import { producerJWKKeyWriterServiceBuilder } from "./producerJWKKeyWriterService.js";
 
 const db = makeDrizzleConnection(config);
-const producerJWKKeyReadModelService =
-  producerJWKKeyReadModelServiceBuilder(db);
-const readModelService = customReadModelServiceBuilder(
-  db,
-  producerJWKKeyReadModelService
-);
+const producerJWKKeyWriterService = producerJWKKeyWriterServiceBuilder(db);
 
 async function processMessage({
   message,
@@ -41,7 +33,9 @@ async function processMessage({
       : generateId<CorrelationId>(),
   });
   await match(decodedMessage)
-    .with({ event_version: 2 }, (msg) => handleMessageV2(msg, readModelService))
+    .with({ event_version: 2 }, (msg) =>
+      handleMessageV2(msg, producerJWKKeyWriterService)
+    )
     .with({ event_version: 1 }, () => Promise.resolve())
     .exhaustive();
