+++ packages/producer-key-readmodel-writer-sql/src/producerKeyConsumerServiceV2.ts
@@ -3,51 +3,53 @@ import {
   AuthorizationEventEnvelopeV2,
   fromProducerKeychainV2,
   genericInternalError,
+  missingKafkaMessageDataError,
   unsafeBrandId,
 } from "pagopa-interop-models";
 import { match, P } from "ts-pattern";
-import { CustomReadModelService } from "./readModelService.js";
+import { ProducerJWKKeyWriterService } from "./producerJWKKeyWriterService.js";
 
 export async function handleMessageV2(
   message: AuthorizationEventEnvelopeV2,
-  producerJWKKeyReadModelService: CustomReadModelService
+  producerJWKKeyWriterService: ProducerJWKKeyWriterService
 ): Promise<void> {
   await match(message)
     .with({ type: "ProducerKeychainKeyAdded" }, async (message) => {
-      const producerKeychain = message.data.producerKeychain
-        ? fromProducerKeychainV2(message.data.producerKeychain)
-        : undefined;
-      const key = producerKeychain?.keys.find(
+      if (!message.data.producerKeychain) {
+        throw missingKafkaMessageDataError("producerKeychain", message.type);
+      }
+
+      const producerKeychain = fromProducerKeychainV2(
+        message.data.producerKeychain
+      );
+
+      const key = producerKeychain.keys.find(
         (key) => key.kid === message.data.kid
       );
       if (!key) {
         throw genericInternalError(
-          `Key not found in producerKeychain: ${producerKeychain?.id}`
+          `Key not found in producerKeychain: ${producerKeychain.id}`
         );
       }
-      const producerKeychainId = producerKeychain?.id;
-      if (!producerKeychainId) {
-        throw genericInternalError("ProducerKeychainId not found");
-      }
 
-      await producerJWKKeyReadModelService.upsertProducerJWKKey(
-        keyToProducerJWKKey(key, producerKeychainId),
+      await producerJWKKeyWriterService.upsertProducerJWKKey(
+        keyToProducerJWKKey(key, producerKeychain.id),
         message.version
       );
     })
     .with({ type: "ProducerKeychainKeyDeleted" }, async (message) => {
       if (!message.data.producerKeychain) {
-        throw genericInternalError("ProducerKeychain not found in message");
+        throw missingKafkaMessageDataError("producerKeychain", message.type);
       }
 
-      await producerJWKKeyReadModelService.deleteProducerJWKKeyByProducerKeychainAndKid(
+      await producerJWKKeyWriterService.deleteProducerJWKKeyByProducerKeychainAndKid(
         unsafeBrandId(message.data.producerKeychain.id),
         message.data.kid,
         message.version
       );
     })
     .with({ type: "ProducerKeychainDeleted" }, async (message) => {
-      await producerJWKKeyReadModelService.deleteProducerJWKKeysByProducerKeychainId(
+      await producerJWKKeyWriterService.deleteProducerJWKKeysByProducerKeychainId(
         unsafeBrandId(message.data.producerKeychainId),
         message.version
       );
