+++ packages/producer-key-readmodel-writer-sql/test/utils.ts
@@ -1,7 +1,14 @@
 import { setupTestContainersVitest } from "pagopa-interop-commons-test";
 import { producerJWKKeyReadModelServiceBuilder } from "pagopa-interop-readmodel";
 import { inject, afterEach } from "vitest";
-import { customReadModelServiceBuilder } from "../src/readModelService.js";
+import { and, eq } from "drizzle-orm";
+import { ProducerKeychainId } from "pagopa-interop-models";
+import {
+  DrizzleReturnType,
+  ProducerJWKKeySQL,
+  producerJwkKeyInReadmodelProducerJwkKey,
+} from "pagopa-interop-readmodel-models";
+import { producerJWKKeyWriterServiceBuilder } from "../src/producerJWKKeyWriterService.js";
 
 export const { cleanup, readModelDB } = await setupTestContainersVitest(
   undefined,
@@ -17,7 +24,26 @@ afterEach(cleanup);
 
 export const producerJWKKeyReadModelService =
   producerJWKKeyReadModelServiceBuilder(readModelDB);
-export const readModelService = customReadModelServiceBuilder(
-  readModelDB,
-  producerJWKKeyReadModelService
-);
+export const producerJWKKeyWriterService =
+  producerJWKKeyWriterServiceBuilder(readModelDB);
+
+export const retrieveProducerJWKKeySQLByProducerKeychainIdAndKid = async (
+  producerKeychainId: ProducerKeychainId,
+  kid: string,
+  db: DrizzleReturnType
+): Promise<ProducerJWKKeySQL | undefined> => {
+  const result = await db
+    .select()
+    .from(producerJwkKeyInReadmodelProducerJwkKey)
+    .where(
+      and(
+        eq(
+          producerJwkKeyInReadmodelProducerJwkKey.producerKeychainId,
+          producerKeychainId
+        ),
+        eq(producerJwkKeyInReadmodelProducerJwkKey.kid, kid)
+      )
+    );
+
+  return result[0];
+};
