+++ packages/readmodel-checker/test/utils.ts
@@ -33,7 +33,6 @@ import {
   producerKeychainReadModelServiceBuilder,
   purposeReadModelServiceBuilder,
   tenantReadModelServiceBuilder,
-  producerJWKKeyReadModelServiceBuilder,
   eserviceTemplateReadModelServiceBuilder,
 } from "pagopa-interop-readmodel";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
@@ -75,8 +74,6 @@ export const producerKeychainReadModelServiceSQL =
   producerKeychainReadModelServiceBuilder(readModelDB);
 export const clientKeysReadModelServiceSQL =
   clientJWKKeyReadModelServiceBuilder(readModelDB);
-export const producerKeychainKeyReadModelServiceSQL =
-  producerJWKKeyReadModelServiceBuilder(readModelDB);
 
 export const addOneEService = async (
   eservice: WithMetadata<EService>
