+++ packages/readmodel/src/producerJWKKeyReadModelService.ts
@@ -1,4 +1,4 @@
-import { eq, and, lte, SQL } from "drizzle-orm";
+import { eq, and, SQL } from "drizzle-orm";
 import {
   genericInternalError,
   ProducerJWKKey,
@@ -9,57 +9,11 @@ import {
   DrizzleReturnType,
   producerJwkKeyInReadmodelProducerJwkKey,
 } from "pagopa-interop-readmodel-models";
-import { splitProducerJWKKeyIntoObjectsSQL } from "./authorization/producerJWKKeySplitters.js";
 import { aggregateProducerJWKKey } from "./authorization/producerJWKKeyAggregators.js";
-import { checkMetadataVersionByFilter } from "./utils.js";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function producerJWKKeyReadModelServiceBuilder(db: DrizzleReturnType) {
   return {
-    async upsertProducerJWKKey(
-      jwkKey: ProducerJWKKey,
-      metadataVersion: number
-    ): Promise<void> {
-      await db.transaction(async (tx) => {
-        const shouldUpsert = await checkMetadataVersionByFilter(
-          tx,
-          producerJwkKeyInReadmodelProducerJwkKey,
-          metadataVersion,
-          and(
-            eq(producerJwkKeyInReadmodelProducerJwkKey.kid, jwkKey.kid),
-            eq(
-              producerJwkKeyInReadmodelProducerJwkKey.producerKeychainId,
-              jwkKey.producerKeychainId
-            )
-          )
-        );
-
-        if (!shouldUpsert) {
-          return;
-        }
-
-        await tx
-          .delete(producerJwkKeyInReadmodelProducerJwkKey)
-          .where(
-            and(
-              eq(
-                producerJwkKeyInReadmodelProducerJwkKey.producerKeychainId,
-                jwkKey.producerKeychainId
-              ),
-              eq(producerJwkKeyInReadmodelProducerJwkKey.kid, jwkKey.kid)
-            )
-          );
-
-        const producerJWKKeySQL = splitProducerJWKKeyIntoObjectsSQL(
-          jwkKey,
-          metadataVersion
-        );
-
-        await tx
-          .insert(producerJwkKeyInReadmodelProducerJwkKey)
-          .values(producerJWKKeySQL);
-      });
-    },
     async getProducerJWKKeyByProducerKeychainIdAndKid(
       producerKeychainId: ProducerKeychainId,
       kid: string
@@ -89,31 +43,8 @@ export function producerJWKKeyReadModelServiceBuilder(db: DrizzleReturnType) {
       }
       return aggregateProducerJWKKey(queryResult[0]);
     },
-
-    async deleteProducerJWKKeyByProducerKeychainAndKid(
-      producerKeychainId: ProducerKeychainId,
-      kid: string,
-      metadataVersion: number
-    ): Promise<void> {
-      await db
-        .delete(producerJwkKeyInReadmodelProducerJwkKey)
-        .where(
-          and(
-            eq(
-              producerJwkKeyInReadmodelProducerJwkKey.producerKeychainId,
-              producerKeychainId
-            ),
-            eq(producerJwkKeyInReadmodelProducerJwkKey.kid, kid),
-            lte(
-              producerJwkKeyInReadmodelProducerJwkKey.metadataVersion,
-              metadataVersion
-            )
-          )
-        );
-    },
   };
 }
-
 export type ProducerJWKKeyReadModelService = ReturnType<
   typeof producerJWKKeyReadModelServiceBuilder
 >;
